장바구니 CRUD 중 C, D 작동

+++ db/migrations/20221206103943_order_details.sql
@@ -3,9 +3,9 @@ CREATE TABLE
   order_details (
     id int not null auto_increment,
     order_id int,
-    price decimal(10,2),
-    total decimal(10,2),
-    payment decimal(10,2),
+    price int,
+    total int,
+    payment int,
     PRIMARY KEY(id),
     FOREIGN KEY (order_id) REFERENCES orders (id)
  );
